refactor: converge controller authority boundaries
